Object restructure paused to update other functionality... lol
